Ayy this thing is getting sweeter every update

+++ headers/universe.h
@@ -25,7 +25,7 @@ struct s_particle
   double charge;
 
   t_particle *bond[7]; /* Points to the bonded particles (max. 7) */
-  int8_t bond_id[7]; /* ID of the particle located at bond[ID] */
+  int64_t bond_id[7]; /* ID of the particle located at bond[ID] */
   double bond_strength[7]; /* Strength of each bond (spring constant) */
   double bond_length[7]; /* Length of the bond */
